+++ CHANGES.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Rails partials allow relative keys, now it is supported by both Parser and Prism scanners
+  and is covered by tests.
 - `check-prism` makes better comparison between parsers for candidate keys.
 
 ## v1.1.2